maange des taches/backend/requirements.txt
flask==2.0.1
pymongo==4.1.1
flask-cors==3.0.10
python-dotenv==0.19.1 